array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,4 @@
+#include <stdio.h>
 #include "function_pointers.h"
 /**
  * array_iterator - executes a function on each elemnt of an array
@@ -8,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
